Add basic setup support and reorganize database code.
Application will check for file inc/config.php.  If this file does not exist, the setup code will be run and this file will be created.
Caveats:
Webserver must have write access to the inc directory during running the setup.
Database input is not sanitized (yet).
If a database of the input name exists, it is overwritten.